SubjectViewModel removed and SubjectListViewModel renamed to SubjectViewModel
