+++ app/src/main/java/android/mahendra/attendancemanager/viewmodels/SubjectViewModel.java
@@ -2,50 +2,84 @@ package android.mahendra.attendancemanager.viewmodels;
 
 import android.mahendra.attendancemanager.models.Subject;
 import android.mahendra.attendancemanager.repositories.SubjectRepository;
-import android.util.Log;
 
-import androidx.databinding.BaseObservable;
+import android.app.Application;
 
+import androidx.lifecycle.AndroidViewModel;
+import androidx.lifecycle.LiveData;
+import androidx.lifecycle.Transformations;
+
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Locale;
 
-public class SubjectViewModel extends BaseObservable {
-    private static final String TAG = "SubjectViewModel";
-    private Subject mSubject;
+public class SubjectViewModel extends AndroidViewModel {
 
-    public void setSubject(Subject subject) {
-        mSubject = subject;
+    private SubjectRepository mSubjectRepository;
+    private LiveData<List<Subject>> mAllSubjects;
+
+    public SubjectViewModel(Application application) {
+        super(application);
+        mSubjectRepository = new SubjectRepository(application);
+        mAllSubjects = mSubjectRepository.getAllSubjects();
     }
 
-    public String getTitle() {
-        return mSubject.getTitle();
+    public LiveData<List<Subject>> getAllSubjects() {
+        return mAllSubjects;
     }
 
-    public String getAttendanceStat() {
+    public void insert(Subject subject) {
+        mSubjectRepository.insert(subject);
+    }
+
+    public void update(Subject subject) {
+        mSubjectRepository.update(subject);
+    }
+
+    private List<String> extractTitles(List<Subject> subjects) {
+        List<String> subjectTitles = new ArrayList<>();
+        for (Subject subject : subjects) {
+            subjectTitles.add(subject.getTitle());
+        }
+        return subjectTitles;
+    }
+
+    public LiveData<List<String>> getSubjectTitles() {
+        return Transformations.map(mAllSubjects, this::extractTitles);
+    }
+
+    public String getTitle(Subject subject) {
+        return subject.getTitle();
+    }
+
+    public String getAttendanceStat(Subject subject) {
         String stat = String.
-                format(Locale.ENGLISH,"%d / %d", mSubject.getAttendedClasses(), mSubject.getTotalClasses());
+                format(Locale.ENGLISH,"%d / %d", subject.getAttendedClasses(), subject.getTotalClasses());
         return stat;
     }
 
-    public String getAttendancePercentage() {
-        if (mSubject.getTotalClasses() == 0) {
+    public String getAttendancePercentage(Subject subject) {
+        if (subject.getTotalClasses() == 0) {
             return "0.0%";
         }
-        double percentage = ((double) mSubject.getAttendedClasses() / mSubject.getTotalClasses()) * 100;
+        double percentage = ((double) subject.getAttendedClasses() / subject.getTotalClasses()) * 100;
         return String.format(Locale.ENGLISH, "%.1f%%", percentage);
     }
 
-    public int getAttendanceProgress() {
-        if (mSubject.getTotalClasses() == 0) {
+    public int getAttendanceProgress(Subject subject) {
+        if (subject.getTotalClasses() == 0) {
             return 0;
         }
-        return (mSubject.getAttendedClasses() * 100) / mSubject.getTotalClasses();
+        return (subject.getAttendedClasses() * 100) / subject.getTotalClasses();
     }
 
-    public void onAttended() {
-        mSubject.incrementClassesAttended();
+    public void onAttended(Subject subject) {
+        subject.incrementClassesAttended();
+        update(subject);
     }
 
-    public void onMissed() {
-        mSubject.incrementClassesMissed();
+    public void onMissed(Subject subject) {
+        subject.incrementClassesMissed();
+        update(subject);
     }
 }
